fix(engine): improve efficiency of authorizations in history
This commit improves the efficiency of authorizations in history.
More precisely, when querying for a historic entity like
"HistoricProcessInstnace", the READ_HISTORY premission needs to be
checked.
The previous implementation checked this permission for every
historic process instance matched by the current query. The
authorization check is constant in time, which is good. However,
since it is executed as a sub-select, a constant factor is added for
each row matched my the query. As a result, this scales linearly with
the number of rows in the history table.

This commit proposes a new implementation which leverages the fact
that the READ_HISTORY permission is defined on the process definition
resource. Meaning: I can either see all instances of a process
definition or none of the instances. As a result, we do not need to
check this permission for every instance but we can check it once for
a process definition. The new algorithm works in the following way:

1) select the distinct set of process definition keys from the table
(this result set will be an order of magnitude of 100s, in some very
rare cases 1000s but not more).
2) for each key in the set, perform the authorization check, and by
this construct the set of all process definition keys for which
READ_HISTORY can be proved.
3) Join this set with the table itself.

Follow up: benchmark

Related to #CAM-5450